Slowly migrating to next.js
We have finished adding the Biography and the Layout of the home page.

Now we have to set up GraphQL/Apollo to parse the blog post markdown files so we can query them to populate the home page. Then we'll need to create a route for the blog posts themselves.